Implements ZMM electrostatic interaction
Zero-multipole summation method (ZMM) is a cutoff-based method to
compute electrostatic interaction for solution and ionic crystal
systems. The main advantage of the method is a good cost / accuracy
tradeoff.

This commit is 1/4 of commit series.
The commit adds the "Zero-multipole" to the electrostatic interaction
and implements it with tabulated kernel.